audit-assistant/requirements.txt
# Core
fastapi>=0.95.0,<0.100.0
uvicorn[standard]>=0.21.0,<0.22.0
python-multipart>=0.0.6,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=1.10.7,<2.0.0
email-validator>=2.0.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.10.0,<2.0.0
psycopg2-binary>=2.9.5,<3.0.0  # For PostgreSQL
aiosqlite>=0.18.0,<0.19.0    # For SQLite

# Async Database
sqlalchemy[asyncio]>=2.0.0,<3.0.0
aiopg>=1.4.0,<2.0.0  # For async PostgreSQL

# Document Processing
python-magic>=0.4.27,<0.5.0
python-magic-bin>=0.4.14,<0.5.0  # Windows only
pypdf2>=3.0.0,<4.0.0
python-docx>=0.8.11,<0.9.0
openpyxl>=3.1.2,<4.0.0
python-pptx>=0.6.21,<0.7.0
unstructured>=0.8.0,<0.9.0  # For advanced document parsing

# Vector Search
numpy>=1.24.0,<2.0.0
sentence-transformers>=2.2.2,<3.0.0
faiss-cpu>=1.7.4,<2.0.0  # For CPU-based similarity search
# faiss-gpu>=1.7.2,<2.0.0  # Uncomment for GPU support

# LLM Providers
openai>=0.27.0,<0.28.0
google-generativeai>=0.1.0,<0.2.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0

# Utilities
python-slugify>=7.0.0,<8.0.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3,<2024.0
tqdm>=4.65.0,<5.0.0
requests>=2.28.2,<3.0.0
httpx>=0.24.0,<0.25.0
python-json-logger>=2.0.6,<3.0.0
python-ulid>=2.2.0,<3.0.0  # For generating unique IDs

# Development
pytest>=7.3.1,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
pytest-cov>=4.0.0,<5.0.0
black>=23.3.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.2.0,<2.0.0
types-python-dateutil>=2.8.19,<3.0.0
types-requests>=2.28.11,<3.0.0
types-pytz>=2023.3.0,<2024.0.0

# Documentation
mkdocs>=1.4.3,<2.0.0
mkdocs-material>=9.1.8,<10.0.0
mkdocstrings[python]>=0.21.2,<0.22.03
